Moved SDKTestWorkflowRule to sdk package
