+++ temporal-sdk/src/test/java/io/temporal/workflow/AsyncActivityCompleteWithErrorTest.java
@@ -23,7 +23,7 @@ import io.temporal.activity.*;
 import io.temporal.activity.ManualActivityCompletionClient;
 import io.temporal.common.RetryOptions;
 import io.temporal.failure.ApplicationFailure;
-import io.temporal.testing.TestWorkflowRule;
+import io.temporal.workflow.shared.SDKTestWorkflowRule;
 import java.time.Duration;
 import java.util.concurrent.ForkJoinPool;
 import org.junit.Assert;
@@ -33,8 +33,8 @@ import org.junit.Test;
 public class AsyncActivityCompleteWithErrorTest {
 
   @Rule
-  public TestWorkflowRule testWorkflowRule =
-      TestWorkflowRule.newBuilder()
+  public SDKTestWorkflowRule testWorkflowRule =
+      SDKTestWorkflowRule.newBuilder()
           .setWorkflowTypes(TestWorkflowImpl.class)
           .setActivityImplementations(new AsyncActivityWithManualCompletion())
           .build();
