+++ temporal-sdk/src/test/java/io/temporal/workflow/AsyncActivityRetry.java
@@ -25,8 +25,8 @@ import io.temporal.client.WorkflowException;
 import io.temporal.client.WorkflowStub;
 import io.temporal.common.RetryOptions;
 import io.temporal.failure.ApplicationFailure;
-import io.temporal.testing.SDKTestWorkflowRule;
 import io.temporal.testing.WorkflowReplayer;
+import io.temporal.workflow.shared.SDKTestWorkflowRule;
 import io.temporal.workflow.shared.TestActivities;
 import io.temporal.workflow.shared.TestWorkflows;
 import java.io.IOException;
