+++ temporal-sdk/src/test/java/io/temporal/workflow/LocalActivityAndQueryTest.java
@@ -22,9 +22,9 @@ package io.temporal.workflow;
 import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.client.WorkflowStub;
-import io.temporal.testing.TestOptions;
 import io.temporal.testing.TestWorkflowRule;
 import io.temporal.workflow.shared.TestActivities;
+import io.temporal.workflow.shared.TestOptions;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
