+++ temporal-sdk/src/test/java/io/temporal/workflow/WorkflowTest.java
@@ -51,11 +51,8 @@ import io.temporal.internal.sync.DeterministicRunnerTest;
 import io.temporal.serviceclient.WorkflowServiceStubs;
 import io.temporal.serviceclient.WorkflowServiceStubsOptions;
 import io.temporal.testing.*;
-import io.temporal.testing.TestOptions;
 import io.temporal.worker.*;
-import io.temporal.workflow.shared.TestActivities;
-import io.temporal.workflow.shared.TestMultiargdsWorkflowFunctions;
-import io.temporal.workflow.shared.TestWorkflows;
+import io.temporal.workflow.shared.*;
 import java.io.*;
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadInfo;
