+++ temporal-sdk/src/test/java/io/temporal/workflow/shared/TestOptions.java
@@ -17,7 +17,7 @@
  *  permissions and limitations under the License.
  */
 
-package io.temporal.testing;
+package io.temporal.workflow.shared;
 
 import io.temporal.activity.ActivityOptions;
 import io.temporal.activity.LocalActivityOptions;
